fix: type of psi_bias

+++ source/source_psi/psi.h
@@ -156,7 +156,7 @@ class Psi
     // current pointer for getting the psi
     mutable T* psi_current = nullptr;
     // psi_current = psi + psi_bias;
-    mutable int psi_bias = 0;
+    mutable size_t psi_bias = 0;
 
     const int* ngk = nullptr;
